Rename libproc_macro to libproc_macro_internal
The name was a bit misleading since libproc_macro is a rust library that
should we should be able to link against some rust code. This cpp
library was the layer below the libproc_macro library, the "real" gcc
libproc_macro act as a translation layer for the internal library.

gcc/rust/ChangeLog:

	* expand/rust-macro-expand.h: Change include directive with the
	new name.
	* expand/rust-proc-macro.h (RUST_PROC_MACRO_H): Likewise.
	* util/rust-hir-map.h: Likewise.
	* util/rust-token-converter.h: Likewise.
	* util/rust-token-converter.cc: Remove useless include
	directive.
	* Make-lang.in: Rename library.

ChangeLog:

	* libgrust/Makefile.am: Change library name.
	* libgrust/Makefile.in: Likewise.
	* libgrust/configure.ac: Likewise.
	* libgrust/configure: Regenerate.
	* libgrust/libproc_macro/Makefile.am: Moved to...
	* libgrust/libproc_macro_internal/Makefile.am: ...here.
	* libgrust/libproc_macro/Makefile.in: Moved to...
	* libgrust/libproc_macro_internal/Makefile.in: ...here.
	* libgrust/libproc_macro/bridge.h: Moved to...
	* libgrust/libproc_macro_internal/bridge.h: ...here.
	* libgrust/libproc_macro/ffistring.cc: Moved to...
	* libgrust/libproc_macro_internal/ffistring.cc: ...here.
	* libgrust/libproc_macro/ffistring.h: Moved to...
	* libgrust/libproc_macro_internal/ffistring.h: ...here.
	* libgrust/libproc_macro/group.cc: Moved to...
	* libgrust/libproc_macro_internal/group.cc: ...here.
	* libgrust/libproc_macro/group.h: Moved to...
	* libgrust/libproc_macro_internal/group.h: ...here.
	* libgrust/libproc_macro/ident.cc: Moved to...
	* libgrust/libproc_macro_internal/ident.cc: ...here.
	* libgrust/libproc_macro/ident.h: Moved to...
	* libgrust/libproc_macro_internal/ident.h: ...here.
	* libgrust/libproc_macro/literal.cc: Moved to...
	* libgrust/libproc_macro_internal/literal.cc: ...here.
	* libgrust/libproc_macro/literal.h: Moved to...
	* libgrust/libproc_macro_internal/literal.h: ...here.
	* libgrust/libproc_macro/proc_macro.cc: Moved to...
	* libgrust/libproc_macro_internal/proc_macro.cc: ...here.
	* libgrust/libproc_macro/proc_macro.h: Moved to...
	* libgrust/libproc_macro_internal/proc_macro.h: ...here.
	* libgrust/libproc_macro/punct.cc: Moved to...
	* libgrust/libproc_macro_internal/punct.cc: ...here.
	* libgrust/libproc_macro/punct.h: Moved to...
	* libgrust/libproc_macro_internal/punct.h: ...here.
	* libgrust/libproc_macro/registration.h: Moved to...
	* libgrust/libproc_macro_internal/registration.h: ...here.
	* libgrust/libproc_macro/rust/bridge.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/bridge.rs: ...here.
	* libgrust/libproc_macro/rust/bridge/ffistring.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/bridge/ffistring.rs: ...here.
	* libgrust/libproc_macro/rust/bridge/group.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/bridge/group.rs: ...here.
	* libgrust/libproc_macro/rust/bridge/ident.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/bridge/ident.rs: ...here.
	* libgrust/libproc_macro/rust/bridge/literal.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/bridge/literal.rs: ...here.
	* libgrust/libproc_macro/rust/bridge/punct.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/bridge/punct.rs: ...here.
	* libgrust/libproc_macro/rust/bridge/span.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/bridge/span.rs: ...here.
	* libgrust/libproc_macro/rust/bridge/token_stream.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/bridge/token_stream.rs: ...here.
	* libgrust/libproc_macro/rust/group.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/group.rs: ...here.
	* libgrust/libproc_macro/rust/ident.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/ident.rs: ...here.
	* libgrust/libproc_macro/rust/lib.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/lib.rs: ...here.
	* libgrust/libproc_macro/rust/literal.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/literal.rs: ...here.
	* libgrust/libproc_macro/rust/punct.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/punct.rs: ...here.
	* libgrust/libproc_macro/rust/span.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/span.rs: ...here.
	* libgrust/libproc_macro/rust/token_stream.rs: Moved to...
	* libgrust/libproc_macro_internal/rust/token_stream.rs: ...here.
	* libgrust/libproc_macro/span.cc: Moved to...
	* libgrust/libproc_macro_internal/span.cc: ...here.
	* libgrust/libproc_macro/span.h: Moved to...
	* libgrust/libproc_macro_internal/span.h: ...here.
	* libgrust/libproc_macro/tokenstream.cc: Moved to...
	* libgrust/libproc_macro_internal/tokenstream.cc: ...here.
	* libgrust/libproc_macro/tokenstream.h: Moved to...
	* libgrust/libproc_macro_internal/tokenstream.h: ...here.
	* libgrust/libproc_macro/tokentree.cc: Moved to...
	* libgrust/libproc_macro_internal/tokentree.cc: ...here.
	* libgrust/libproc_macro/tokentree.h: Moved to...
	* libgrust/libproc_macro_internal/tokentree.h: ...here.
