+++ libgrust/libproc_macro_internal/Makefile.am
@@ -48,7 +48,7 @@ AM_MAKEFLAGS = \
         "TARGET_LIB_PATH=$(TARGET_LIB_PATH)"
 
 
-TARGETLIB = ./libproc_macro.a
+TARGETLIB = ./libproc_macro_internal.a
 LIBOBJS = @LIBOBJS@
 objext = @OBJEXT@
