+++ libgrust/libproc_macro_internal/rust/bridge/ffistring.rs
